requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cliff>=2.2.0